pyproject.toml
[project]
name = "medmnist-project"
version = "0.1.0"
requires-python = ">=3.11,<3.15"
dependencies = [
    "fastapi==0.128.0",
    "numpy==2.4.1",
    "pillow==12.1.0",
    "python-multipart==0.0.21",
    "uvicorn==0.40.0"
]

[project.optional-dependencies]
train = [
    "ipykernel==7.1.0",
    "matplotlib==3.10.8",
    "medmnist==3.0.2",
    "requests==2.32.5",
    "torch==2.9.1",
    "torchvision==0.24.1"
]